assignments/messy-migration/requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
flask-bcrypt==1.0.1
flask-limiter==3.5.0
python-dotenv==1.0.0
pytest==7.4.0